+++ sequana/scripts/compressor.py
@@ -19,16 +19,13 @@
 """Standalone dedicated to taxonomic content (kraken)"""
 import os
 import sys
-import shutil
-import sys
-import platform
 from optparse import OptionParser
 import argparse
 
-import tempfile
-
 from sequana.scripts.tools import SequanaOptions
+from sequana import misc
 from sequana import Module, SequanaConfig
+
 from easydev import SmartFormatter, TempFile
 
 
@@ -137,7 +134,7 @@ def main(args=None):
         raise ValueError('The number of jobs is limited to 20. You can ' +
             'force this limit by using --bypass-job-limit')
 
-    if platform.uname().node.startswith("tars-"):
+    if misc.on_cluster("tars-") and options.unlock is False:
         if options.cluster is None:
             raise ValueError("You are on TARS (Institut Pasteur). You " +
                 " must use --cluster option to provide the scheduler " +
@@ -172,7 +169,6 @@ Must be one of fastq, fastq.gz, fastq.bz2 or fastq.dsrc""")
         cfg.config.compressor.threads = options.threads
         cfg._update_yaml()
         cfg.save(filename=temp.name)
-
 
         # The Snakefile can stay in its original place:
         rule = module.path + os.sep +  "compressor.rules"