+++ test/scripts/test_sequana_pipeline.py
@@ -1,22 +1,14 @@
 from sequana.scripts import main
 from sequana import sequana_data
-
+import pytest
+from unittest.mock import patch
 
 prog = "sequana"
 
-"""    def teardown_class(klass):
-        try:os.remove('quality.rules')
-        except:pass
-        try:os.remove('config.yaml')
-        except:pass
 
-        import shutil
-        try:shutil.rmtree("Hm2_test")
-        except:pass
+def test_main():
+    main.main([prog])
 
-        try:shutil.rmtree("report")
-        except:pass
-"""
 
 def test_version():
     main.main([prog, '--version'])
@@ -50,25 +42,108 @@ def test_help():
     else:
         raise Exception
 
-def _test_info():
-    main.main([prog, '--info', "quality"])
+def test_info(mocker):
+    def func(*args, **kwargs):
+        pass
+    with patch('easydev.onweb', func ):
+        main.main([prog, '--info', "quality_control"])
+
+# This mock does not work somehow, it opens the web page... is it in conflict
+# with the test_info ?
+def _test_issue(mocker):
+    def func(*args, **kwargs):
+        pass
+    with patch('easydev.onweb', func ):
+        main.main([prog, '--issue'])
 
 def test_show_pipelines():
     main.main([prog, '--show-pipelines'])
 
-#def test_mutually_exclusive():
-#    try:
-#        main.main([prog, '--pipeline', 'quality', '--info'])
-#        assert False
-#    except:
-#        assert True
+def test_mutually_exclusive():
+    try:
+        main.main([prog, '--pipeline', 'quality_control', '--info', "quality_control"])
+        assert False
+    except SystemExit:
+        assert True
 
-def test_input():
+def test_input(tmpdir):
+    directory = tmpdir.mkdir("analysis")
+    name = directory.__str__()
+
     file1 = sequana_data('Hm2_GTGAAA_L005_R1_001.fastq.gz', 'data')
     file2 = sequana_data('Hm2_GTGAAA_L005_R2_001.fastq.gz', 'data')
     main.main([prog, "--pipeline", "quality_control", "--file1", 
-              file1, "--file2", file2, "--no-adapters" , "--force"])
+              file1, "--file2", file2, "--no-adapters" , "--force",
+              "--working-directory", name])
+
+    try:
+        main.main([prog, "--pipeline", "quality_control", "--file1", 
+              file1+"dummy", "--no-adapters" , "--force",
+              "--working-directory", name])
+        assert False
+    except ValueError:
+        assert True
+    try:
+        main.main([prog, "--pipeline", "quality_control", "--file1", 
+              file1, "--no-adapters" , "--force", "--file2",
+              file2+"duly","--working-directory", name])
+        assert False
+    except ValueError:
+        assert True
+    try:
+        main.main([prog, "--pipeline", "quality_control", "--file1", 
+              file1, "--no-adapters" , "--force", "--file2",
+              file2,"--working-directory", name, "--kraken", "dummy"])
+        assert False
+    except ValueError:
+        assert True
+    try:
+        main.main([prog, "--pipeline", "quality_control", "--file1", 
+              file1, "--no-adapters" , "--force", "--file2",
+              file2,"--working-directory", name, "--kraken", "dummy"])
+        assert False
+    except ValueError:
+        assert True
 
 
+def test_without_cluster_config(tmpdir):
+    directory = tmpdir.mkdir("analysis")
+    name = directory.__str__()
+    file1 = sequana_data('Hm2_GTGAAA_L005_R1_001.fastq.gz', 'data')
+    main.main([prog, "--pipeline", "rnaseq", "--file1", file1, "--snakemake-cluster", 
+        '"sbatch --mem={cluster.ram}"', "--ignore-cluster-config", "--force",
+        "--working-directory", name])
+    with open("%s/runme.sh" %  name) as fh: 
+        assert "cluster_config" not in fh.read()
+
+# Fails on travis
+def _test_with_cluster_config(tmpdir):
+    directory = tmpdir.mkdir("analysis2")
+    name = directory.__str__()
+    file1 = sequana_data('Hm2_GTGAAA_L005_R1_001.fastq.gz', 'data')
+    main.main([prog, "--pipeline", "rnaseq", "--file1", file1, "--snakemake-cluster", 
+        '"sbatch --mem={cluster.ram}"',"--force",
+        "--working-directory", name])
+    with open("%s/runme.sh" %  name) as fh: 
+        assert "cluster_config" in fh.read()
 
 
+def test_get_config():
+
+    main.main([prog,  "--pipeline", "quality_control",
+        "--get-config"])
+    import os
+    os.remove("config.yaml")
+
+
+def test_config_params():
+    file1 = sequana_data('Hm2_GTGAAA_L005_R1_001.fastq.gz', 'data')
+    main.main([prog, "--pipeline", "quality_control", "--force",  "--file1", file1,
+        "--no-adapters",  "--config-params",  "bwa_mem_phix:threads:4"])
+
+    try:
+        main.main([prog, "--pipeline", "quality_control", "--force",  "--file1", file1,
+            "--no-adapters",  "--config-params",  "bwa_mem_phix"])
+        assert False
+    except:
+        assert True