+++ doc/Changelog.rst
@@ -3,12 +3,55 @@ Changelog
 
 .. contents::
 
+0.1.21 - Feb 2017
+-------------------
+
+* NEWS:
+
+    - add sequana_debug_level function at top level to switch verbosity of
+      informative messages (default is WARNING).
+    - add pacbio module  #351
+    - quality control pipeline: atropos can be used in place of cutadapt #346
+
+* CHANGES:
+
+   - Running Median is 10 times faster #345
+   - sequana_coverage:  (1) --file1 alone was not working (2) automatically copy
+     cluster-config in working directory and update runme.sh accordingly #342
+   - sequana standalone:
+       - handles cluster_config Snakemake option
+       - add error message when adapter name is incorrect
+   - sequanix: the help dialog is now created inside designer and has a proper
+     scrollable browser dialog. cluster_config Snakemake option is also handle.
+   - Remove galleria JS lib and related files (htmltools)
+   - sequana_coverage: add --logging-level option
+
+* BUG:
+
+    - Fix #352 : allow gc window size to be even (warning is shown and +1 to
+      window size)
+    - Fix # 354: cutadapt report that was mixing up R1/R2 trimming in the images.
+    - --output-directory in sequana_coverage was failing 
+    - in coverage, centralness was buggy (regression) and use number of ROIs
+      instead of the total base length #347
+    - Fix multi_report summary for single end case #349
+
+0.1.20 - Feb 2017
+---------------------
+
+* CHANGES: 
+
+    - remove pyquickhelper dependencies and add a simple rest2html function in
+      misc module.
+
 0.1.19 - Feb 2017
 -----------------------
 
 
 * CHANGES:
 
+    - misc module: factorise on_cluster() function used in compressor scripts to
+        be used in other tools such as sequanix
     - compressor: limits max number of jobs to 20 (can be bypass manually),
       prevent run on TARS if snakemake-cluster not provided. 
     - rules:
@@ -27,6 +70,8 @@ Changelog
 
 * NEWS:
 
+    - Sequanix stable version
+    - add TrueSeq adaptors
     - add lazy import mechanism to speed up the time to import sequana, which 
       speeds up the --help in the standalone